+++ src/gpsl/variable.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 #[derive(Clone, PartialEq, Debug, Serialize, Deserialize)]
 pub enum Variable {
-    Number { value: usize },
+    Number { value: i64 },
     Text { value: String },
     Return { value: Box<Variable> },
     None {},
